refactor: ChatContext랑 chatContextsh 통합
